feature/#12/patch Schedules Date API 구현완료 1. 오늘자 기준으로 이후의 schedules date의 데이터들을 삭제 2. 그리고 schedules common 에 등록된 새로운 날짜/주기 기준으로 계산하면서, 오늘 날짜 이후의 데이터만 새로 넣어줌

+++ app/main/service/schedules_common.py
@@ -13,7 +13,6 @@ from app.main.model.schedules_date import Schedules_date
 from app.main.model.users import Users
 from ..config import jwt_key, jwt_alg
 import re
-from sqlalchemy.sql.functions import now
 
 def post_schedules_common(data):
   """ Post Common information of alarm"""
@@ -141,40 +140,52 @@ def edit_schedules_date(data):
         print(startdate, enddate, cycle)
         #여기까지 post와 동일
         
-        currdate = now.strftime('%Y-%m-%d %H:%M:%S') #오늘자 기준으로 for loop 돌리기
+        #오늘날짜
+        now = datetime.datetime.now()
+        today = now.strftime('%Y-%m-%d') #오늘자 기준으로 for loop 돌리기
+        currdate = datetime.datetime.strptime(today, '%Y-%m-%d')
+        print(currdate)
         # 1. 우선 DB내에서 오늘이후의 schedules_date 지우기
         while currdate <=enddate:
-          saved_schedule = Schedules_date.query.filter_by(alarmdate=currdate).first()
-          db.session.delete(saved_schedule)
-          db.session.commit()
-          currdate = currdate + datetime.timedelta(days=cycle)
+          saved_schedule = Schedules_date.query.filter(and_(Schedules_date.alarmdate==currdate, Schedules_date.schedules_common_id==schedules_common_id)).first()
+          if saved_schedule:
+            print('delete:',saved_schedule)
+            db.session.delete(saved_schedule)
+          currdate = currdate + datetime.timedelta(days=1)
 
+        #그리고 오늘 이후 부터 새로운 cycle/time 적용해서 일정 재등록
+        currdate_new = datetime.datetime.strptime(today, '%Y-%m-%d')
+        print(currdate_new)
         
-        # while currdate <= enddate:
-        #   #이를 schedules_date 테이블에 넣어주기
-        #   new_schedules_date = Schedules_date(
-        #     alarmdate = currdate,
-        #     time = time,
-        #     check = 0,
-        #     user_id = user_id,
-        #     schedules_common_id = schedules_common_id
-        #   )
-        #   db.session.add(new_schedules_date)
-        #   db.session.commit()
-        #   currdate = currdate + datetime.timedelta(days=cycle)
+        #schedule_common에 등록된 startdate로 부터 주기대로 계산하다가 오늘이 넘는 날짜부터 새로 등록하기
+        currdate_for_cal = startdate
+        while currdate_for_cal <= enddate:
+          if currdate_for_cal >= currdate_new:
+            print('save:',currdate_for_cal)
+            #이를 schedules_date 테이블에 넣어주기
+            new_schedules_date = Schedules_date(
+              alarmdate = currdate_for_cal,
+              time = time,
+              check = 0,
+              user_id = user_id,
+              schedules_common_id = schedules_common_id
+            )
+            db.session.add(new_schedules_date)
+            db.session.commit()
+          currdate_for_cal = currdate_for_cal + datetime.timedelta(days=cycle)
+          #print(currdate_for_cal)
 
-        # #response는 medicine_id와 schedules_common_id
-        # results = {
-        #   'medicine_id': medicine_id,
-        #   'schedules_common_id': schedules_common_id
-        # }
+        #response는 medicine_id와 schedules_common_id
+        results = {
+          'schedules_common_id': schedules_common_id
+        }
 
-        # response_object = {
-        #   'status': 'OK',
-        #   'message': 'Successfully post schedules date.',
-        #   'results': results
-        # }
-        # return response_object, 200
+        response_object = {
+          'status': 'OK',
+          'message': 'Successfully post schedules date.',
+          'results': results
+        }
+        return response_object, 200
 
     except Exception as e:
       print(e)
